event_generator/requirements.txt
adal==1.2.2
asn1crypto==0.24.0
azure-common==1.1.23
azure-eventhub==1.3.1
azure-storage-blob==1.5.0
azure-storage-common==1.4.2
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
idna==2.8
isodate==0.6.0
msrest==0.6.9
msrestazure==0.6.1
oauthlib==3.1.0
pycparser==2.19
PyJWT==1.7.1
python-dateutil==2.8.0
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
uamqp==1.2.2
urllib3==1.25.3
